refact:merge InputView, OutputView int View
