--- src/main/java/racingcar/InputView.java
@@ -1,11 +1,0 @@
-package racingcar;
-
-import java.util.List;
-
-public interface InputView {
-    int getNumberOfCar();
-
-    int getNumberOfRacing();
-
-    List<String> getCarNames();
-}
