+++ src/main/java/racingcar/RacingCarMain.java
@@ -7,28 +7,27 @@ public class RacingCarMain {
     public static void main(String[] args) {
         final int MOVE_THRESHOLD = 4;
 
-        InputView inputView = new Step4InputView();
+        View view = new Step4View();
 
-        List<String> carNames = inputView.getCarNames();
+        List<String> carNames = view.getCarNames();
         List<Car> cars = new ArrayList<Car>();
         carNames.forEach(name -> cars.add(new SimpleCar(name)));
         CarCollection carCollection = new CarCollection(cars);
 
-        int numberOfRacing = inputView.getNumberOfRacing();
+        int numberOfRacing = view.getNumberOfRacing();
 
-        OutputView outputView = new Step4OutputView();
         Racing racing = new Racing(new RandomMovementCondition(MOVE_THRESHOLD), carCollection, numberOfRacing);
         racing.addEndOfSingleRacingListener(new Observer<Integer>() {
             @Override
             public void observe(Integer currentRacing) {
-                outputView.printSingleRacingResult(cars);
+                view.printSingleRacingResult(cars);
             }
         });
 
-        outputView.printTitle();
+        view.printTitle();
         racing.racing();
 
         List<String> winnersName = carCollection.getWinnersName();
-        outputView.printWinners(winnersName);
+        view.printWinners(winnersName);
     }
 }
