--- src/main/java/racingcar/Step3OutputView.java
@@ -1,30 +1,0 @@
-package racingcar;
-
-import java.util.List;
-
-public class Step3OutputView implements OutputView {
-    final String FORWARD = "-";
-
-    @Override
-    public void printTitle() {
-        final String TITLE = "실행 결과";
-        System.out.println();
-        System.out.println(TITLE);
-    }
-
-    @Override
-    public void printSingleRacingResult(List<Car> cars) {
-        cars.forEach(car -> {
-            int distance = car.getDistance();
-            while (distance-- != 0) {
-                System.out.print(FORWARD);
-            }
-            System.out.print('\n');
-        });
-        System.out.print('\n');
-    }
-
-    @Override
-    public void printWinners(List<String> winnersName) {
-    }
-}
