+++ src/main/java/racingcar/Step3View.java
@@ -4,7 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Scanner;
 
-public class Step3InputView implements InputView {
+public class Step3View implements View {
+    final String FORWARD = "-";
+
     private Scanner scanner = new Scanner(System.in);
 
     @Override
@@ -27,5 +29,28 @@ public class Step3InputView implements InputView {
     public List<String> getCarNames() {
         return new ArrayList<String>() {
         };
+    }
+
+    @Override
+    public void printTitle() {
+        final String TITLE = "실행 결과";
+        System.out.println();
+        System.out.println(TITLE);
+    }
+
+    @Override
+    public void printSingleRacingResult(List<Car> cars) {
+        cars.forEach(car -> {
+            int distance = car.getDistance();
+            while (distance-- != 0) {
+                System.out.print(FORWARD);
+            }
+            System.out.print('\n');
+        });
+        System.out.print('\n');
+    }
+
+    @Override
+    public void printWinners(List<String> winnersName) {
     }
 }
