+++ src/main/java/racingcar/Step4View.java
@@ -4,7 +4,10 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Scanner;
 
-public class Step4InputView implements InputView {
+public class Step4View implements View {
+    final String FORWARD = "-";
+    final String CAR_NAME_POSTFIX = " : ";
+
     private Scanner scanner = new Scanner(System.in);
 
     @Override
@@ -30,5 +33,37 @@ public class Step4InputView implements InputView {
         System.out.println(QUESTION_FOR_GET_NUMBER_OF_CARS);
         int numberOfCars = scanner.nextInt();
         return numberOfCars;
+    }
+
+    @Override
+    public void printTitle() {
+        final String TITLE = "실행 결과";
+        System.out.println();
+        System.out.println(TITLE);
+    }
+
+    @Override
+    public void printSingleRacingResult(List<Car> cars) {
+        cars.forEach(car -> {
+            int distance = car.getDistance();
+            System.out.print(car.getName() + CAR_NAME_POSTFIX);
+            while (distance-- != 0) {
+                System.out.print(FORWARD);
+            }
+            System.out.print('\n');
+        });
+        System.out.print('\n');
+    }
+
+    @Override
+    public void printWinners(List<String> winnersName) {
+        if (winnersName.isEmpty()) return;
+
+        System.out.print(winnersName.get(0));
+        for (int i = 1; i < winnersName.size(); i++) {
+            System.out.print(", ");
+            System.out.print(winnersName.get(i));
+        }
+        System.out.print("가 최종 우승했습니다.\n");
     }
 }
