+++ src/main/java/racingcar/View.java
@@ -1,0 +1,17 @@
+package racingcar;
+
+import java.util.List;
+
+public interface View {
+    int getNumberOfCar();
+
+    int getNumberOfRacing();
+
+    List<String> getCarNames();
+
+    void printTitle();
+
+    void printSingleRacingResult(List<Car> cars);
+
+    void printWinners(List<String> winnersName);
+}
